moved performance code to an application code instead of test
