+++ performance/README.md
@@ -3,7 +3,7 @@
 ## DataSetGenerator 
 
 ```shell
- mvn exec:java -Dscope=test -Dexec.mainClass="com.clickhouse.benchmark.data.ClickHouseDataTypesShort" -input <table_fields.sql> -rows <number_of_rows>
+ mvn exec:java -Dscope=test -Dexec.mainClass="com.clickhouse.com.clickhouse.benchmark.data.ClickHouseDataTypesShort" -input <table_fields.sql> -rows <number_of_rows>
 ```
 
 
@@ -11,6 +11,6 @@
 
 with custom dataset
 ```shell
-mvn test-compile exec:java -Dexec.classpathScope=test -Dexec.mainClass="com.clickhouse.benchmark.BenchmarkRunner" -Dexec.args="--dataset=file://dataset_1741150759025.csv"
+mvn test-compile exec:java -Dexec.classpathScope=test -Dexec.mainClass="com.clickhouse.com.clickhouse.benchmark.BenchmarkRunner" -Dexec.args="--dataset=file://dataset_1741150759025.csv"
 ```
