+++ performance/src/main/java/com/clickhouse/benchmark/clients/ConcurrentInsertClient.java
@@ -1,7 +1,10 @@
 package com.clickhouse.benchmark.clients;
 
 import com.clickhouse.benchmark.BenchmarkRunner;
+import com.clickhouse.benchmark.data.DataSet;
+import com.clickhouse.client.ClickHouseClient;
 import com.clickhouse.client.ClickHouseResponse;
+import com.clickhouse.client.api.Client;
 import com.clickhouse.client.api.data_formats.RowBinaryFormatWriter;
 import com.clickhouse.client.api.insert.InsertResponse;
 import com.clickhouse.client.api.insert.InsertSettings;
@@ -13,50 +16,92 @@ import com.clickhouse.data.ClickHouseSerializer;
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.Level;
 import org.openjdk.jmh.annotations.Scope;
+import org.openjdk.jmh.annotations.Setup;
 import org.openjdk.jmh.annotations.State;
 import org.openjdk.jmh.annotations.TearDown;
+import org.openjdk.jmh.annotations.Threads;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.List;
 
 import static com.clickhouse.benchmark.TestEnvironment.getServer;
-
+@Threads(3)
 @State(Scope.Benchmark)
-public class InsertClient extends BenchmarkBase {
-    private static final Logger LOGGER = LoggerFactory.getLogger(InsertClient.class);
-
-    @TearDown(Level.Invocation)
-    public void verifyRowsInsertedAndCleanup(DataState dataState) {
-        boolean success;
-        int count = 0;
-        do {
-            success = verifyCount(dataState.tableNameEmpty, dataState.dataSet.getSize());
-            if (!success) {
-                LOGGER.warn("Retrying to verify rows inserted");
-                try {
-                    Thread.sleep(2500);
-                } catch (InterruptedException e) {
-                    LOGGER.error("Error: ", e);
-                }
-            }
-        } while (!success && count++ < 10);
-        if (!success) {
-            LOGGER.error("Failed to verify rows inserted");
-            throw new RuntimeException("Failed to verify rows inserted");
+public class ConcurrentInsertClient extends BenchmarkBase {
+    private static final Logger LOGGER = LoggerFactory.getLogger(ConcurrentInsertClient.class);
+    private static ClickHouseClient clientV1Shared;
+    private static Client clientV2Shared;
+    @Setup(Level.Trial)
+    public void setUpIteration() {
+        clientV1Shared = getClientV1();
+        clientV2Shared = getClientV2();
+    }
+    @TearDown(Level.Trial)
+    public void tearDownIteration() {
+        if (clientV1Shared != null) {
+            clientV1Shared.close();
+            clientV1Shared = null;
         }
-        truncateTable(dataState.tableNameEmpty);
+        if (clientV2Shared != null) {
+            clientV2Shared.close();
+            clientV2Shared = null;
+        }
+    }
+    @State(Scope.Thread)
+    public static class ThreadLocalState {
+        public String createTableName() {
+            String name = Thread.currentThread().getName();
+            int index = name.lastIndexOf("-");
+            String id = name.substring(index + 1);
+            return String.format("%s_%s", "concurrent_data_empty", id);
+        }
+        @Setup(Level.Trial)
+        public void setup() {
+            DataSet dataSet = DataState.getDataSet();
+            String tableName = createTableName();
+            LOGGER.warn("setup create table name: " + tableName);
+            // create table
+            String createTableString = dataSet.getCreateTableString(tableName);
+            runAndSyncQuery(createTableString, tableName);
+        }
+
+        @TearDown(Level.Invocation)
+        public void verifyRowsInsertedAndCleanup(DataState dataState) {
+            String tableName = createTableName();
+            boolean success;
+            LOGGER.warn("TearDown: " + tableName);
+            int count = 0;
+            do {
+                success = verifyCount(tableName, dataState.dataSet.getSize());
+                if (!success) {
+                    LOGGER.warn("Retrying to verify rows inserted");
+                    try {
+                        Thread.sleep(2500);
+                    } catch (InterruptedException e) {
+                        LOGGER.error("Error: ", e);
+                    }
+                }
+            } while (!success && count++ < 10);
+            if (!success) {
+                LOGGER.error("Failed to verify rows inserted");
+                throw new RuntimeException("Failed to verify rows inserted");
+            }
+            truncateTable(tableName);
+        }
     }
 
+
     @Benchmark
-    public void insertV1(DataState dataState) {
+    public void insertV1(DataState dataState, ThreadLocalState threadLocalState) {
+        String tableName = threadLocalState.createTableName();
         try {
             ClickHouseFormat format = dataState.dataSet.getFormat();
-            try (ClickHouseResponse response = clientV1.read(getServer())
+            try (ClickHouseResponse response = clientV1Shared.read(getServer())
                     .write()
                     .option(ClickHouseClientOption.ASYNC, false)
                     .format(format)
-                    .query(BenchmarkRunner.getInsertQuery(dataState.tableNameEmpty))
+                    .query(BenchmarkBase.getInsertQuery(tableName))
                     .data(out -> {
                         for (byte[] bytes: dataState.dataSet.getBytesList(format)) {
                             out.write(bytes);
@@ -70,10 +115,12 @@ public class InsertClient extends BenchmarkBase {
     }
 
     @Benchmark
-    public void insertV2(DataState dataState) {
+    public void insertV2(DataState dataState, ThreadLocalState threadLocalState) {
+        String tableName = threadLocalState.createTableName();
+        LOGGER.warn("insertV2: " + tableName);
         try {
             ClickHouseFormat format = dataState.dataSet.getFormat();
-            try (InsertResponse response = clientV2.insert(dataState.tableNameEmpty, out -> {
+            try (InsertResponse response = clientV2Shared.insert(tableName, out -> {
                 for (byte[] bytes: dataState.dataSet.getBytesList(format)) {
                     out.write(bytes);
                 }
@@ -87,15 +134,15 @@ public class InsertClient extends BenchmarkBase {
     }
 
     @Benchmark
-    public void insertV1Compressed(DataState dataState) {
+    public void insertV1Compressed(DataState dataState, ThreadLocalState threadLocalState) {
         try {
             ClickHouseFormat format = dataState.dataSet.getFormat();
-            try (ClickHouseResponse response = clientV1.read(getServer())
+            try (ClickHouseResponse response = clientV1Shared.read(getServer())
                     .write()
                     .option(ClickHouseClientOption.ASYNC, false)
                     .option(ClickHouseClientOption.DECOMPRESS, true)
                     .format(format)
-                    .query(BenchmarkRunner.getInsertQuery(dataState.tableNameEmpty))
+                    .query(BenchmarkBase.getInsertQuery(threadLocalState.createTableName()))
                     .data(out -> {
                         for (byte[] bytes: dataState.dataSet.getBytesList(format)) {
                             out.write(bytes);
@@ -107,12 +154,11 @@ public class InsertClient extends BenchmarkBase {
             LOGGER.error("Error: ", e);
         }
     }
-
     @Benchmark
-    public void insertV2Compressed(DataState dataState) {
+    public void insertV2Compressed(DataState dataState, ThreadLocalState threadLocalState) {
         try {
             ClickHouseFormat format = dataState.dataSet.getFormat();
-            try (InsertResponse response = clientV2.insert(dataState.tableNameEmpty, out -> {
+            try (InsertResponse response = clientV2Shared.insert(threadLocalState.createTableName(), out -> {
                 for (byte[] bytes: dataState.dataSet.getBytesList(format)) {
                     out.write(bytes);
                 }
@@ -127,17 +173,17 @@ public class InsertClient extends BenchmarkBase {
     }
 
     @Benchmark
-    public void insertV1RowBinary(DataState dataState) {
+    public void insertV1RowBinary(DataState dataState, ThreadLocalState threadLocalState) {
         try {
             ClickHouseFormat format = ClickHouseFormat.RowBinary;
-            try (ClickHouseResponse response = clientV1.read(getServer())
+            try (ClickHouseResponse response = clientV1Shared.read(getServer())
                     .write()
                     .option(ClickHouseClientOption.ASYNC, false)
                     .format(format)
-                    .query(BenchmarkRunner.getInsertQuery(dataState.tableNameEmpty))
+                    .query(BenchmarkBase.getInsertQuery(threadLocalState.createTableName()))
                     .data(out -> {
                         ClickHouseDataProcessor p = dataState.dataSet.getClickHouseDataProcessor();
-                        ClickHouseSerializer[] serializers = p.getSerializers(clientV1.getConfig(), p.getColumns());
+                        ClickHouseSerializer[] serializers = p.getSerializers(clientV1Shared.getConfig(), p.getColumns());
                         for (ClickHouseRecord record : dataState.dataSet.getClickHouseRecords()) {
                             for (int i = 0; i < serializers.length; i++) {
                                 serializers[i].serialize(record.getValue(i), out);
@@ -153,9 +199,9 @@ public class InsertClient extends BenchmarkBase {
     }
 
     @Benchmark
-    public void insertV2RowBinary(DataState dataState) {
+    public void insertV2RowBinary(DataState dataState, ThreadLocalState threadLocalState) {
         try {
-            try (InsertResponse response = clientV2.insert(dataState.tableNameEmpty, out -> {
+            try (InsertResponse response = clientV2Shared.insert(threadLocalState.createTableName(), out -> {
                 RowBinaryFormatWriter w = new RowBinaryFormatWriter(out, dataState.dataSet.getSchema(), ClickHouseFormat.RowBinary);
                 for (List<Object> row : dataState.dataSet.getRowsOrdered()) {
                     int index = 1;
@@ -174,5 +220,4 @@ public class InsertClient extends BenchmarkBase {
             LOGGER.error("Error: ", e);
         }
     }
-
 }
